common-repo/services/Eci-order-service/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.2
pyyaml==6.0.1
mysqlclient>=2.2
requests>=2.32
pika>=1.3